--- src/components/pages/StudentDashboard/StudentViewRumble/StudentRumblePages/PastRumbleDetails/index.ts
@@ -1,1 +1,0 @@
-export { default as PastRumbleDetails } from './PastRumbleDetailsContainer';
