+++ src/components/pages/StudentDashboard/StudentViewRumble/StudentRumblePages/ViewSubmission/ViewSubmissionContainer.tsx
@@ -3,21 +3,16 @@ import { useRecoilValue } from 'recoil';
 import { Students, Submissions } from '../../../../../../api';
 import { auth, current } from '../../../../../../state';
 import { CouldNotLoad, Loader } from '../../../../../common';
-import RenderPastRumbleDetails from './RenderPastRumbleDetails';
+import RenderPastRumbleDetails from './RenderViewSubmission';
 
 const PastRumbleDetailsContainer = (): React.ReactElement => {
   const rumble = useRecoilValue(current.rumble);
   const user = useRecoilValue(auth.user);
   const section = useRecoilValue(current.section);
-  // Removed []
   const [submission, setSubmission] = useState<Submissions.ISubItem>();
   const [error, setError] = useState<null | string>(null);
 
   useEffect(() => {
-    // console.log('rumbleId & userId: ', {
-    //   rumbleId: rumble?.id,
-    //   userId: user?.id,
-    // });
     if (rumble && user && !submission) {
       Students.getSubForRumble(rumble.id, user.id)
         .then((res) => {
