+++ src/components/pages/StudentDashboard/StudentViewRumble/StudentRumblePages/ViewSubmission/index.ts
@@ -1,0 +1,1 @@
+export { default as PastRumbleDetails } from './ViewSubmissionContainer';
