+++ src/components/pages/StudentDashboard/StudentViewRumble/StudentRumblePages/index.ts
@@ -1,4 +1,4 @@
-export { PastRumbleDetails } from './PastRumbleDetails';
+export { PastRumbleDetails } from './ViewSubmission';
 export { PeerFeedbackPage } from './PeerFeedbackPage';
 export { default as RumbleComplete } from './RumbleComplete';
 export { StudentSubmissionPage } from './StudentSubmissionPage';
